+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # DF Architect
 
 ## Release 3.1.0 (2021-11-02)
+- NEW: Added specialized code to deal with PF2e's breaking and heavy-handed overhaul of the FoundryVTT canvas Lighting Layer.
 - FIX: Image scaling caused by Pixel Resolution Scaling. Will now temporarily disable PRS while capturing the canvas.
 - FIX: Wall Snapping issue when hovering over an existing wall causing that wall to move instead.
 - FIX: Conflict with Theatre Inserts.